+++ examples/example_plugin/doc/gen_doc.ml
@@ -1,0 +1,3 @@
+
+let _ = Elpi.API.BuiltIn.document_file ~header:"% Generated"
+  Elpi_example_plugin.Rocq_elpi_example_plugin.builtins